refactor(project): Use lerna for monorepo, add dockerfile
