+++ Dockerfile
@@ -1,0 +1,9 @@
+FROM node:latest
+
+COPY . /app
+WORKDIR /app
+RUN npm install
+RUN npm run bootstrap
+RUN npm run build
+
+CMD ["npm", "run", "start"]
