+++ lerna.json
@@ -1,0 +1,5 @@
+{
+    "$schema": "node_modules/lerna/schemas/lerna-schema.json",
+    "useWorkspaces": true,
+    "version": "0.1.0"
+}
