+++ package.json
@@ -1,22 +1,19 @@
 {
-  "name": "search-helper",
-  "version": "1.0.0",
-  "engines": {
-    "node": "18.12.1"
-  },
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon app.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "mongodb": "^4.12.0"
-  }
+    "name": "search-helper",
+    "private": true,
+    "version": "1.0.0",
+    "scripts": {
+        "lerna": "lerna",
+        "start": "lerna run start --stream",
+        "build": "lerna run build --stream",
+        "test": "lerna run test --stream",
+        "dev": "lerna run dev --stream",
+        "bootstrap": "lerna bootstrap"
+    },
+    "workspaces": [
+        "packages/*"
+    ],
+    "devDependencies": {
+        "lerna": "^6.1.0"
+    }
 }
