+++ packages/client/.gitignore
@@ -1,1 +1,2 @@
+build/
 node_modules/
