+++ packages/client/public/favicon.ico
