+++ packages/client/public/manifest.json
