+++ packages/client/src/modules/cookieHandler.js
