+++ packages/client/src/sass/_SearchBar.scss
