+++ packages/server/app.js
@@ -1,0 +1,20 @@
+const express = require("express");
+const cors = require("cors");
+
+const { PORT, CLIENT_APPLICATION, checkEnvVars } = require("./env_vars");
+require("dotenv").config({ path: "./.env" });
+checkEnvVars();
+
+const api = require("./api");
+
+const app = express();
+
+app.use(cors());
+app.use(express.json());
+
+app.get("/", express.static(CLIENT_APPLICATION));
+app.use("/api", api);
+
+app.listen(PORT, () => {
+    console.log(`listening on port ${PORT}`);
+});
