+++ packages/server/package.json
@@ -1,0 +1,21 @@
+{
+    "name": "search-helper-server",
+    "description": "The backend for the search helper app",
+    "version": "0.1.0",
+    "main": "app.js",
+    "scripts": {
+        "start": "node .",
+        "dev": "nodemon app.js",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "license": "ISC",
+    "dependencies": {
+        "cors": "^2.8.5",
+        "dotenv": "^16.0.3",
+        "express": "^4.18.2",
+        "mongodb": "^4.12.0"
+    },
+    "devDependencies": {
+        "nodemon": "^2.0.20"
+    }
+}
